Exclude code in C/C++ 'assert' from CCN
Consider a simple use case:

assert((conditional_expression) && "More description...");

This line adds 1 + CCN(conditional_expression)
if the assertion is counted as a part of a function.

Since assertions are mostly used
to check (pre, post)-conditions,
to explicitly state assumptions,
and to catch programmer errors,
they are not expected to be a part of the flow of a function.
Moreover, assertions with side-effects is a very bad idea.
The function should stay the same
if all the assertions are removed (in release mode, for example).