Use Python3 division for calculation of averages
In some places, it seems that the results of calculations
are expected to be floats,
but integer division of Python2 is used.
In other places, explicit cast to float is made to get the float result.
Python3 division approach for integers helps cleanup these places
and make them consistent.

If Lizard relies on integer division approach of Python2,
which doesn't seem to be the case,
the code must be corrected with '//'.