Implement counting of Nested Control Structures
The tests and implementation are borrowed from the Nesting Depth code;
however, the counting of the nested control structures
is implemented with stack for simplicity,
which can be inefficient (worth measuring!).

Note that this implementation does not provide
user facing front-end controls to generate warnings.
It only prints the metric in the summary page.

The implementation is only for C-like languages
with C++ being the main target.

The implementation is only tested with simple code snippets
and not verified against production code.