Fix `nesting_level` for brace-less ctrl structures
The nesting structure extension
is running before the assignment of the nesting level for tokens.
The solution is to increment the nesting level
right at the end of control-flow structure declaration
but before its body.

In addition,
the token is yielded before being processed
to simulate running the extension after the code reader,
but this is somewhat hacky.

This bug (surprisingly) has been a feature!
The fix breaks the gotcha braceless nested if-then-else structures.
The current implementation of the metric looks wrong overall.

Fixes #157.