Fixing bug when python has multiple line for function declaration due to formatting (#319)
* Fixing bug when python has multiple line for function declaration due to formatting

Currently there's bug in python Lizard when getting LOC and CCN when func declaration has multiple lines. It will only count the function declaration lines but not the implementation. This is because the original code uses indentation to check function ending while for multi-line declaration, the last line will start with ")" with no indent.

Example:
```
def f1(a,
    b
):
  # function implementation for 10 lines
```

Issue in https://github.com/terryyin/lizard/issues/292

* Add unit test for pull request #319

* Fix testPython.py